% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libfuzzerr-package.R
\docType{package}
\name{libfuzzerr-package}
\alias{libfuzzerr}
\alias{libfuzzerr-package}
\title{libfuzzerr: Expements in using 'libfuzzerr'}
\description{
Expements in using 'libfuzzerr'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mpadge/libfuzzerr}
  \item Report bugs at \url{https://github.com/mpadge/libfuzzerr/issues}
}

}
\author{
\strong{Maintainer}: Mark Padgham \email{mark.padgham@email.com}

}
\keyword{internal}
